Code cleanup & also added folders for easier project navigation
